// src/lib/utils/enums.ts
export const enum Sources {
  Stable = 'stable',
  Main = 'main',
  Rpc = 'rpc',
  Collection = 'collection',
  Builders = 'builders',
  Voice = 'voice'
}

export enum DocTypes {
  Class = 'class',
  Event = 'event',
  Interface = 'interface',
  Method = 'method',
  Param = 'param',
  Prop = 'prop',
  Typedef = 'typedef'
}